Project/requirements.txt
--upgrade setuptools
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
django-filter
firebase-admin